+++ .dockerignore
@@ -1,2 +1,21 @@
-node_modules
+# .vscode/
+# .dockerignore   # Ignore the ignore
+# node_modules    # Ignore local node_modules folder
+# npm-debug.log   # Debug files
+# Dockerfile      # The dockerfile
+# .git            # The git history
+# .gitignore      #
+# .npmrc          # If accessing a private npm repository here will be the token used, so ignore to prevent leaking
+# .env-*          # Any other environment that we don't want to include
+# .github         # Using github actions ?
+# *.md            # Any 
+# .eslintrc.js
+# .prettierrc.
+# docker-nodeJS-CheatSheet.pdf
+# Makefile
+.vscode/
+node_modules/
 npm-debug.log
+dist/
+graphql/
+test/
