+++ .prettierrc
@@ -1,4 +1,10 @@
 {
   "singleQuote": true,
   "trailingComma": "all"
+  
 }
+
+// "semi": true,
+  // "useTabs": true,
+  // "tabWidth": 4,
+  // "arrowParens": "always"
