--- dockerfile.dev
@@ -1,14 +1,0 @@
-FROM node:18-alpine
-
-WORKDIR /app/
-
-COPY ./package.json ./
-COPY prisma ./prisma/
-
-RUN npm install
-
-
-
-COPY . .
-
-CMD [ "npm", "run", "start:migrate"]
