+++ package.json
@@ -10,7 +10,7 @@
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
     "start:dev": "nest start --watch",
-    "start:migrate": "prisma migrate dev && nest start --watch",
+    "start:migrate": "prisma migrate dev && prisma db seed && npm run start:dev",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
@@ -23,7 +23,9 @@
     "test:e2e:watch": "dotenv -e .env.test -- jest --watch --config ./test/jest-e2e.json --maxWorkers=1",
     "test:e2e": "dotenv -e .env.test -- jest --config ./test/jest-e2e.json --maxWorkers=1",
     "migrate-test:postgres": "dotenv -e .env.test -- npx prisma migrate deploy",
-    "studio-test:postgres": "dotenv -e .env.test -- npx prisma studio"
+    "studio-test:postgres": "dotenv -e .env.test -- npx prisma studio",
+    "prisma:generate": "npx prisma generate",
+    "prisma:migrate": "npx prisma migrate dev"
   },
   "dependencies": {
     "@golevelup/ts-jest": "^0.3.4",
@@ -92,7 +94,8 @@
     "testEnvironment": "node"
   },
   "prisma": {
-    "seed": "ts-node prisma/seed.ts",
-    "seed-test": "ts-node prisma/seed.ts"
+    "seed": "ts-node src/database/prisma/seed.ts",
+    "seed-test": "ts-node prisma/seed.ts",
+    "schema": "src/database/schema.prisma"
   }
 }
