+++ src/database/prisma/libs/prisma.ts
@@ -1,0 +1,2 @@
+import { PrismaClient } from '@prisma/client';
+export const prisma = new PrismaClient();
