+++ src/database/prisma/seed.ts
@@ -1,0 +1,772 @@
+import { prisma } from './libs/prisma';
+import { reportData, userData } from './data';
+
+// 0e5c503c-ddea-4cfc-a789-4c913273ace4
+// e5d7f1d2-48b0-46f7-9eee-b5e0be377c8d
+async function main() {
+  try {
+    await prisma.user.createMany({
+      data: [
+        // { email: 'martin@email.io', password: 'password' },
+        // { email: 'ben@email.io', password: 'password' },
+        { email: 'saliba@email.io', password: 'password' },
+        { email: 'gabriel@email.io', password: 'password' },
+      ],
+    });
+    // await prisma.report.createMany({
+    //   data: [
+    //     {
+    //       make: 'MINI',
+    //       model: 'COOPER',
+    //       price: 13490,
+    //       color: 'Pepper White',
+    //       transmission: 'MANUAL',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2016,
+    //       mileage: 111944,
+    //     },
+    //     {
+    //       make: 'KIA',
+    //       model: 'Forte',
+    //       price: 11500,
+    //       color: 'Titanium Bronze',
+    //       transmission: 'MANUAL',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 51708,
+    //     },
+    //     {
+    //       make: 'FIAT',
+    //       model: '500',
+    //       price: 10799,
+    //       color: 'Nero Puro',
+    //       transmission: 'MANUAL',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2016,
+    //       mileage: 35162,
+    //     },
+    //     {
+    //       make: 'Honda',
+    //       model: 'Civic Coupe',
+    //       price: 17999,
+    //       color: 'White',
+    //       transmission: 'MANUAL',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 95973,
+    //     },
+    //     {
+    //       make: 'Subaru',
+    //       model: 'WRX',
+    //       price: 16750,
+    //       color: 'Pure Red',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 97126,
+    //     },
+    //     {
+    //       make: 'Subaru',
+    //       model: 'Crosstrek',
+    //       price: 27477,
+    //       color: 'Red',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 22899,
+    //     },
+    //     {
+    //       make: 'Subaru',
+    //       model: 'Forester',
+    //       price: 18999,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2020,
+    //       mileage: 22899,
+    //     },
+    //     {
+    //       make: 'Mitsubishi',
+    //       model: 'Mirage G4',
+    //       price: 13545,
+    //       color: 'Mercury Gray',
+    //       transmission: 'MANUAL',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 10828,
+    //     },
+    //     {
+    //       make: 'Mitsubishi',
+    //       model: 'Eclipse Cross',
+    //       price: 17299,
+    //       color: 'Alloy Silver Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 54057,
+    //     },
+    //     {
+    //       make: 'Mitsubishi',
+    //       model: 'Outlander Sport',
+    //       price: 16889,
+    //       color: 'Diamond White Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 55012,
+    //     },
+    //     {
+    //       make: 'Mitsubishi',
+    //       model: 'Lancer GT',
+    //       price: 11495,
+    //       color: 'Wicked White Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2014,
+    //       mileage: 92827,
+    //     },
+    //     {
+    //       make: 'Mazda',
+    //       model: 'CX-30',
+    //       price: 19825,
+    //       color: 'White Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2020,
+    //       mileage: 25085,
+    //     },
+    //     {
+    //       make: 'Mazda',
+    //       model: 'MAZDA3',
+    //       price: 25647,
+    //       color: 'White Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 44260,
+    //     },
+    //     {
+    //       make: 'Mazda',
+    //       model: 'CX-30',
+    //       price: 19538,
+    //       color: 'Crystal Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2020,
+    //       mileage: 44260,
+    //     },
+    //     {
+    //       make: 'Mazda',
+    //       model: 'MAZDA6',
+    //       price: 19995,
+    //       color: 'Gray Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2020,
+    //       mileage: 42242,
+    //     },
+    //     {
+    //       make: 'Land Rover',
+    //       model: 'Defender',
+    //       price: 97777,
+    //       color: 'Black',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2023,
+    //       mileage: 3144,
+    //     },
+    //     {
+    //       make: 'Land Rover',
+    //       model: 'Range Rover Velar',
+    //       price: 49770,
+    //       color: 'Fuji White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 19314,
+    //     },
+    //     {
+    //       make: 'Land Rover',
+    //       model: 'Range Rover',
+    //       price: 75770,
+    //       color: 'Black Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2020,
+    //       mileage: 29135,
+    //     },
+    //     {
+    //       make: 'Land Rover',
+    //       model: 'Discovery Sport',
+    //       price: 41800,
+    //       color: 'Black Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2022,
+    //       mileage: 8942,
+    //     },
+    //     {
+    //       make: 'Land Rover',
+    //       model: 'Range Rover Evoque',
+    //       price: 19989,
+    //       color: 'White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 53948,
+    //     },
+    //     {
+    //       make: 'BMW',
+    //       model: 'X3',
+    //       price: 42985,
+    //       color: 'Jet Black',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2022,
+    //       mileage: 11188,
+    //     },
+    //     {
+    //       make: 'BMW',
+    //       model: '3 Series',
+    //       price: 29991,
+    //       color: 'Glacier Silver',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 41945,
+    //     },
+    //     {
+    //       make: 'BMW',
+    //       model: 'I3 ',
+    //       price: 28512,
+    //       color: 'Capparis White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 28512,
+    //     },
+    //     {
+    //       make: 'BMW',
+    //       model: 'X5',
+    //       price: 38168,
+    //       color: 'Mineral White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2020,
+    //       mileage: 54707,
+    //     },
+    //     {
+    //       make: 'BMW',
+    //       model: '4 Series',
+    //       price: 17995,
+    //       color: 'White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2015,
+    //       mileage: 74697,
+    //     },
+    //     {
+    //       make: 'Chevrolet',
+    //       model: 'Equinox',
+    //       price: 16965,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2018,
+    //       mileage: 50892,
+    //     },
+    //     {
+    //       make: 'Chevrolet',
+    //       model: 'Bolt EV',
+    //       price: 24988,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2019,
+    //       mileage: 19286,
+    //     },
+    //     {
+    //       make: 'Chevrolet',
+    //       model: 'Spark EV',
+    //       price: 12403,
+    //       color: 'Summit White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2016,
+    //       mileage: 49067,
+    //     },
+    //     {
+    //       make: 'Chevrolet',
+    //       model: 'Camaro',
+    //       price: 33991,
+    //       color: 'Gray Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 28250,
+    //     },
+    //     {
+    //       make: 'Honda',
+    //       model: 'CR-V',
+    //       price: 25891,
+    //       color: 'Silver Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 41758,
+    //     },
+    //     {
+    //       make: 'Honda',
+    //       model: 'Accord',
+    //       price: 26994,
+    //       color: 'Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 12966,
+    //     },
+    //     {
+    //       make: 'Honda',
+    //       model: 'Accord Hybrid',
+    //       price: 33593,
+    //       color: 'White Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2021,
+    //       mileage: 19900,
+    //     },
+    //     {
+    //       make: 'Honda',
+    //       model: 'Accord Hybrid',
+    //       price: 33593,
+    //       color: 'White Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2021,
+    //       mileage: 19900,
+    //     },
+    //     {
+    //       make: 'Honda',
+    //       model: 'CR-Z',
+    //       price: 16995,
+    //       color: 'Blue',
+    //       transmission: 'MANUAL',
+    //       fuel_type: 'HYBRID',
+    //       year: 2014,
+    //       mileage: 47981,
+    //     },
+    //     {
+    //       make: 'Ford',
+    //       model: 'Mustang Mach-E',
+    //       price: 59488,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2021,
+    //       mileage: 9022,
+    //     },
+    //     {
+    //       make: 'Ford',
+    //       model: 'Fusion SE',
+    //       price: 14995,
+    //       color: 'White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2016,
+    //       mileage: 68543,
+    //     },
+    //     {
+    //       make: 'Ford',
+    //       model: 'EcoSport',
+    //       price: 16989,
+    //       color: 'Luxe Yellow',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 14423,
+    //     },
+    //     {
+    //       make: 'Ford',
+    //       model: 'Fusion Hybrid',
+    //       price: 18594,
+    //       color: 'Black',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 44022,
+    //     },
+    //     {
+    //       make: 'Ford',
+    //       model: 'Escape Hybrid',
+    //       price: 26888,
+    //       color: 'White Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2020,
+    //       mileage: 45880,
+    //     },
+    //     {
+    //       make: 'Ford',
+    //       model: 'C-Max',
+    //       price: 17900,
+    //       color: 'Brown',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2018,
+    //       mileage: 42566,
+    //     },
+    //     {
+    //       make: 'Hyundai',
+    //       model: 'Elantra',
+    //       price: 17500,
+    //       color: 'Brown',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2018,
+    //       mileage: 20185,
+    //     },
+    //     {
+    //       make: 'Hyundai',
+    //       model: 'Tucson',
+    //       price: 17870,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 63511,
+    //     },
+    //     {
+    //       make: 'Hyundai',
+    //       model: 'Sonata',
+    //       price: 25199,
+    //       color: 'Quartz White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 20025,
+    //     },
+    //     {
+    //       make: 'Hyundai',
+    //       model: 'Santa Fe',
+    //       price: 12598,
+    //       color: 'Sparkling Silver',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2018,
+    //       mileage: 106208,
+    //     },
+    //     {
+    //       make: 'Kia',
+    //       model: 'Rio',
+    //       price: 14498,
+    //       color: 'Champagne',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2018,
+    //       mileage: 49389,
+    //     },
+    //     {
+    //       make: 'Kia',
+    //       model: 'Sportage',
+    //       price: 16232,
+    //       color: 'White Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 80116,
+    //     },
+    //     {
+    //       make: 'Kia',
+    //       model: 'Soul EV',
+    //       price: 9980,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2019,
+    //       mileage: 99190,
+    //     },
+    //     {
+    //       make: 'Kia',
+    //       model: 'Niro',
+    //       price: 21495,
+    //       color: 'White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2019,
+    //       mileage: 65508,
+    //     },
+    //     {
+    //       make: 'Kia',
+    //       model: 'Optima Hybrid',
+    //       price: 24991,
+    //       color: 'Blue',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2020,
+    //       mileage: 24993,
+    //     },
+    //     {
+    //       make: 'Kia',
+    //       model: 'Forte',
+    //       price: 15900,
+    //       color: 'White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 40082,
+    //     },
+    //     {
+    //       make: 'Lexus',
+    //       model: 'NX',
+    //       price: 27231,
+    //       color: 'White Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 49338,
+    //     },
+    //     {
+    //       make: 'Lexus',
+    //       model: 'GX',
+    //       price: 29995,
+    //       color: 'White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 49338,
+    //     },
+    //     {
+    //       make: 'Lexus',
+    //       model: 'UX',
+    //       price: 30664,
+    //       color: 'Blue',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 22621,
+    //     },
+    //     {
+    //       make: 'Lexus',
+    //       model: 'CT',
+    //       price: 17641,
+    //       color: 'Gray Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2017,
+    //       mileage: 113479,
+    //     },
+    //     {
+    //       make: 'Lexus',
+    //       model: 'IS',
+    //       price: 28988,
+    //       color: 'White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 32085,
+    //     },
+    //     {
+    //       make: 'Mercedes-Benz',
+    //       model: 'CLA-Class',
+    //       price: 19495,
+    //       color: 'Red',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2018,
+    //       mileage: 59132,
+    //     },
+    //     {
+    //       make: 'Mercedes-Benz',
+    //       model: 'B-CLASS',
+    //       price: 19990,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2017,
+    //       mileage: 24423,
+    //     },
+    //     {
+    //       make: 'Nissan',
+    //       model: 'Leaf SV',
+    //       price: 17743,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2018,
+    //       mileage: 45765,
+    //     },
+    //     {
+    //       make: 'Nissan',
+    //       model: 'Altima',
+    //       price: 18400,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2020,
+    //       mileage: 63745,
+    //     },
+    //     {
+    //       make: 'Nissan',
+    //       model: 'Sentra',
+    //       price: 19800,
+    //       color: 'Silver Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2022,
+    //       mileage: 12571,
+    //     },
+    //     {
+    //       make: 'Nissan',
+    //       model: 'Pathfinder',
+    //       price: 18199,
+    //       color: 'Glacier White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 63828,
+    //     },
+    //     {
+    //       make: 'Nissan',
+    //       model: 'Murano',
+    //       price: 21998,
+    //       color: 'Black',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2018,
+    //       mileage: 34708,
+    //     },
+    //     {
+    //       make: 'Nissan',
+    //       model: 'Rogue',
+    //       price: 17110,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 81375,
+    //     },
+    //     {
+    //       make: 'Nissan',
+    //       model: 'Rogue',
+    //       price: 19988,
+    //       color: 'Gray',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 38202,
+    //     },
+    //     {
+    //       make: 'Toyota',
+    //       model: 'Prius Prime',
+    //       price: 17988,
+    //       color: 'Hypersonic Red',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2017,
+    //       mileage: 99768,
+    //     },
+    //     {
+    //       make: 'Toyota',
+    //       model: 'Camry',
+    //       price: 18788,
+    //       color: 'Blue',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 48202,
+    //     },
+    //     {
+    //       make: 'Toyota',
+    //       model: 'Corolla',
+    //       price: 18890,
+    //       color: 'Sand Pearl',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2021,
+    //       mileage: 28202,
+    //     },
+    //     {
+    //       make: 'Toyota',
+    //       model: 'Avalon Hybrid',
+    //       price: 24599,
+    //       color: 'Creme Brulee',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'HYBRID',
+    //       year: 2016,
+    //       mileage: 34270,
+    //     },
+    //     {
+    //       make: 'Toyota',
+    //       model: 'Tacoma',
+    //       price: 9995,
+    //       color: 'Black',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2013,
+    //       mileage: 191289,
+    //     },
+    //     {
+    //       make: 'Volkswagen',
+    //       model: 'Jetta',
+    //       price: 15888,
+    //       color: 'Pure White',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 56474,
+    //     },
+    //     {
+    //       make: 'Volkswagen',
+    //       model: 'E-GOLF',
+    //       price: 21450,
+    //       color: 'Black',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2019,
+    //       mileage: 56474,
+    //     },
+    //     {
+    //       make: 'Volkswagen',
+    //       model: 'Tiguan',
+    //       price: 18499,
+    //       color: 'Green Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 26628,
+    //     },
+    //     {
+    //       make: 'Volkswagen',
+    //       model: 'Golf GTI',
+    //       price: 23400,
+    //       color: 'Blue',
+    //       transmission: 'MANUAL',
+    //       fuel_type: 'GASOLINE',
+    //       year: 2019,
+    //       mileage: 26628,
+    //     },
+    //     {
+    //       make: 'Volkswagen',
+    //       model: 'ID.4',
+    //       price: 36749,
+    //       color: 'Silver Metallic',
+    //       transmission: 'AUTOMATIC',
+    //       fuel_type: 'ELECTRIC',
+    //       year: 2021,
+    //       mileage: 32926,
+    //     },
+    //   ],
+    // });
+  } catch (e) {
+    console.log(e);
+  }
+}
+main()
+  .then(async () => {
+    await prisma.$disconnect();
+  })
+  .catch(async (e) => {
+    console.log(e);
+    await prisma.$disconnect();
+    process.exit(1);
+  });
