+++ src/database/schema.prisma
@@ -1,0 +1,46 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+generator client {
+  provider      = "prisma-client-js"
+  binaryTargets = ["native", "linux-musl"]
+}
+
+model User {
+  id       String   @id @default(uuid())
+  email    String   @unique
+  password String
+  admin    Boolean  @default(true)
+  reports  Report[]
+}
+
+model Report {
+  id           String  @id @default(uuid())
+  price        Int
+  make         String
+  model        String
+  color        String
+  transmission String
+  fuel_type    String
+  year         Int
+  mileage      Int
+  approved     Boolean @default(false)
+  author       User?   @relation(fields: [authorId], references: [id], onDelete: Cascade, onUpdate: Cascade)
+  authorId     String?
+}
+
+// enum TRANSMISSION {
+//   MANUAL
+//   AUTOMATIC
+// }
+
+// enum FUELTYPE {
+//   GASOLINE
+//   HYBRID
+//   ELECTRIC
+// }
